backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
itsdangerous
requests
SQLAlchemy
pandas